Changed the default state of the connectivity to online rather than offline. This will prevent the app from very briefly showing the red offline warning every time the app is opened before detecting connection and disappearing.
On the other hand, not showing the warning for a very brief time (~0.2s) when the app is opened and there is no connectivity is not an issue and is a more logical behaviour.

Better solution to this may be implemented in the future like incorporating the detection of connectivity inside the HomeBloc for instance.